requirements.txt
discord.py
random-cat==1.0.1